se arreglaron los metodos de tipodocumento controller

+++ app/Http/Controllers/gestion_documento/TipoDocumentoController.php
@@ -27,12 +27,10 @@ class TipoDocumentoController extends Controller
     {
         $data = $request->all();
         $tipoDocumentos = TipoDocumento::with($data['relations'] ?? $this->relations)
-        ->where(function($query){
-            QueryUtil::whereCompany($query);
-        });
-        if (isset($data['tituloDocumento'])) {
-            $tipoDocumentos->where('tituloDocumento', 'like', '%' . $data['tituloDocumento'] . '%');
-        }
+            ->where(function ($query) {
+                QueryUtil::whereCompany($query);
+            });
+        $tipoDocumentos = QueryUtil::whereLike($tipoDocumentos, $data, 'tituloDocumento');
         return response()->json($tipoDocumentos->get($data['columns'] ?? $this->columns));
     }
 
@@ -44,14 +42,13 @@ class TipoDocumentoController extends Controller
      */
     public function store(Request $request)
     {
-        $this->authorize('create',TipoDocumento::class);
+        $this->authorize('create', TipoDocumento::class);
         $data = $request->all();
-        $tipoDocumentoData = QueryUtil::createWithCompany($data['tipoDocumento']);
+        $tipoDocumentoData = QueryUtil::createWithCompany($data["tipoDocumento"]);
         $tipoDocumento = TipoDocumento::create($tipoDocumentoData);
-        $tipoDocumento->save();
         $idTipoDocumento = $tipoDocumento->id;
         $tipoDocumento = TipoDocumento::with($data['relations'] ?? $this->relations);
-        return response()->json($tipoDocumento -> find($idTipoDocumento,$data['columns'] ?? $this->columns), 201);
+        return response()->json($tipoDocumento->find($idTipoDocumento, $data['columns'] ?? $this->columns), 201);
     }
 
     /**
@@ -60,10 +57,18 @@ class TipoDocumentoController extends Controller
      * @param  int $id
      * @return \Illuminate\Http\Response
      */
-    public function show(int $id)
+    public function show(Request $request,int $id)
     {
-        $tipoDocumento = TipoDocumento::find($id);
+        $data = $request->all();
 
+        $tipoDocumento = TipoDocumento::with($data['relations'] ?? $this->relations)
+            ->where(function ($query) {
+                QueryUtil::whereCompany($query);
+            })->findOrFail($id,$data['columns'] ?? $this->columns);
+
+        if (!$tipoDocumento) {
+            return response()->json(['Tipo documento no encontrado'], 404);
+        }
         return response()->json($tipoDocumento);
     }
 
@@ -74,14 +79,22 @@ class TipoDocumentoController extends Controller
      * @param  int $id
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, int $id)
+    public function update(Request $request,int $id)
     {
+        // $this->authorize('update', TipoDocumento::class);
         $data = $request->all();
-        $tipoDocumento = TipoDocumento::findOrFail($id);
-        $tipoDocumento->fill($data);
-        $tipoDocumento->save();
 
-        return response()->json($tipoDocumento);
+        return response() -> json($id);
+        // $tipoDocumento = TipoDocumento::with($data['relations'] ?? $this->relations)
+        //     ->where(function ($query) {
+        //         QueryUtil::whereCompany($query);
+        //     })->find($id, $data['columns'] ?? $this->columns);
+
+        // $tipoDocumentoData = QueryUtil::createWithCompany($data['tipoDocumento']);
+        // $tipoDocumento->fill($tipoDocumentoData);
+        // $tipoDocumento->save();
+
+        // return response()->json($tipoDocumento);
     }
 
     /**
@@ -92,7 +105,11 @@ class TipoDocumentoController extends Controller
      */
     public function destroy(int $id)
     {
-        $tipoDocumento = TipoDocumento::findOrFail($id);
+        $this->authorize('delete', TipoDocumento::class);
+        $tipoDocumento = TipoDocumento::query()
+            ->where(function ($query) {
+                QueryUtil::whereCompany($query);
+            })->findOrFail($id);
         $tipoDocumento->delete();
 
         return response()->json([], 204);
